+++ common/scripts/definitions/objects/obstacles.ts
@@ -163,7 +163,7 @@ Obstacles.insert(
     }),
     {
         idString:"barrel",
-        health:100,
+        health:110,
         hitbox:new CircleHitbox2D(v2.new(0,0),0.57),
         scale:{
             destroy:0.6