+++ changelog.md
@@ -1,3 +1,4 @@
+# Surgemd Pre-Beta-4
 ## News
 __                                                                                                      __
 ### Mains
@@ -31,6 +32,7 @@ __                                                                              
 * Explosions Physics
 * Metal Sound
 * Change Damage Indicator Color
+* Now The Default Emote Key Change To V
 __                                                                                                      __
 ## Fixes
 * Fix Backpacks Pickup(Now The You Can Take A Backpack Without DESTROY de previous)